Refactor config and bitbucket hook
