+++ lib/config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  PORT: process.env.PORT || 5000,
+  BOT_API_KEY: process.env.BOT_API_KEY,
+  TELEGRAM_API_KEY: process.env.TELEGRAM_API_KEY,
+  TELEGRAM_API_URL: "https://api.telegram.org",
+  TELEGRAM_CHAT_ID: process.env.TELEGRAM_CHAT_ID
+};
